requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.4
    # via
    #   -r requirements.in
    #   aiohttp-jinja2
aiohttp-jinja2==1.5.1
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
anyio==3.6.2
    # via
    #   starlette
    #   watchfiles
async-timeout==4.0.2
    # via
    #   aiohttp
    #   redis
attrs==22.2.0
    # via
    #   aiohttp
    #   pytest
bcrypt==4.0.1
    # via -r requirements.in
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via aiohttp
click==8.1.3
    # via uvicorn
configobj==5.0.8
    # via
    #   secrets
    #   systematic
cryptography==39.0.2
    # via pyopenssl
exceptiongroup==1.1.0
    # via pytest
fastapi==0.92.0
    # via
    #   -r requirements.in
    #   fastapi-cache2
fastapi-cache2[redis]==0.2.1
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
future==0.18.3
    # via systematic
h11==0.14.0
    # via uvicorn
httptools==0.5.0
    # via uvicorn
humanize==4.6.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   yarl
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   aiohttp-jinja2
markupsafe==2.1.2
    # via jinja2
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
packaging==23.0
    # via pytest
pendulum==2.1.2
    # via fastapi-cache2
pluggy==1.0.0
    # via pytest
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.2.8
    # via python-ldap
pycparser==2.21
    # via cffi
pydantic==1.10.5
    # via fastapi
pyopenssl==23.0.0
    # via -r requirements.in
pytest==7.2.2
    # via pytest-asyncio
pytest-asyncio==0.20.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via pendulum
python-dotenv==1.0.0
    # via uvicorn
python-gnupg==0.5.0
    # via secrets
python-ldap==3.4.3
    # via secrets
pytzdata==2020.1
    # via pendulum
pyyaml==6.0
    # via uvicorn
redis==4.5.1
    # via fastapi-cache2
secrets==1.0.2
    # via -r requirements.in
six==1.16.0
    # via
    #   configobj
    #   python-dateutil
sniffio==1.3.0
    # via anyio
starlette==0.25.0
    # via fastapi
systematic==4.8.7
    # via secrets
tomli==2.0.1
    # via pytest
typing-extensions==4.5.0
    # via pydantic
uvicorn[standard]==0.20.0
    # via
    #   -r requirements.in
    #   fastapi-cache2
uvloop==0.17.0
    # via uvicorn
watchfiles==0.18.1
    # via uvicorn
websockets==10.4
    # via uvicorn
yarl==1.8.2
    # via aiohttp
